refactor(eth): split chain spec out of beacon_types
beacon_types grew to hold fork tables, blob params and gindex helpers
alongside SSZ definitions. Move the spec/fork logic into chain_spec so
each file has a single responsibility.
